Update stage_pi.py
Starting point for trying to finally address this.

We can possibly not provide the referencing mode to the stage, query whether the stage is referenced or not (servo too), and prompt the user if it is not.

The tkinter messagebox is not perfect. Again, given that it is in the model, it is operating in a different process than the view. As such, it is not top level. If I trigger the messagebox from the controller, then the model thread is no longer blocked and the remainder of the stage initialization will be executed.

Co-Authored-By: Kevin Dean <42547789+AdvancedImagingUTSW@users.noreply.github.com>